+++ grafanalib/tests/test_zabbix.py
@@ -15,7 +15,7 @@ def test_serialization_zabbix_target():
     """Serializing a graph doesn't explode."""
     graph = G.Graph(
         title="CPU Usage",
-        dataSource="Zabbix data source",
+        datasource="Zabbix data source",
         targets=[
             Z.zabbixMetricTarget(
                 group="Zabbix Group",
@@ -31,6 +31,7 @@ def test_serialization_zabbix_target():
             G.YAxis(format=G.SHORT_FORMAT, label="CPU seconds / second"),
             G.YAxis(format=G.SHORT_FORMAT),
         ),
+        gridPos=G.GridPos(h=8, w=12, x=0, y=0)
     )
     stream = StringIO()
     _gen.write_dashboard(graph, stream)